Apply overrides at lighthouse binary level
