+++ consensus/types/src/chain_spec.rs
@@ -605,6 +605,12 @@ pub struct Config {
     #[serde(default)]
     pub preset_base: String,
 
+    #[serde(with = "eth2_serde_utils::quoted_u256")]
+    pub terminal_total_difficulty: Uint256,
+    pub terminal_block_hash: Hash256,
+    #[serde(with = "eth2_serde_utils::quoted_u64")]
+    pub terminal_block_hash_activation_epoch: Epoch,
+
     #[serde(with = "eth2_serde_utils::quoted_u64")]
     min_genesis_active_validator_count: u64,
     #[serde(with = "eth2_serde_utils::quoted_u64")]
@@ -707,6 +713,10 @@ impl Config {
         Self {
             preset_base: T::spec_name().to_string(),
 
+            terminal_total_difficulty: spec.terminal_total_difficulty,
+            terminal_block_hash: spec.terminal_block_hash,
+            terminal_block_hash_activation_epoch: spec.terminal_block_hash_activation_epoch,
+
             min_genesis_active_validator_count: spec.min_genesis_active_validator_count,
             min_genesis_time: spec.min_genesis_time,
             genesis_fork_version: spec.genesis_fork_version,
@@ -750,6 +760,9 @@ impl Config {
         // Pattern match here to avoid missing any fields.
         let &Config {
             ref preset_base,
+            terminal_total_difficulty,
+            terminal_block_hash,
+            terminal_block_hash_activation_epoch,
             min_genesis_active_validator_count,
             min_genesis_time,
             genesis_fork_version,
@@ -799,6 +812,9 @@ impl Config {
             deposit_chain_id,
             deposit_network_id,
             deposit_contract_address,
+            terminal_total_difficulty,
+            terminal_block_hash,
+            terminal_block_hash_activation_epoch,
             ..chain_spec.clone()
         })
     }
